isolate doctors and speciality files
