+++ lib/features/home/ui/widgets/doctors_list/doctors_list_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../data/models/specialization_response_model.dart';
+import '../../../data/models/specialization_response_model.dart';
 import 'doctors_list_view_item.dart';
 
 class DoctorsListView extends StatelessWidget {
