+++ lib/features/home/ui/widgets/doctors_list/doctors_list_view_item.dart
@@ -2,9 +2,9 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../core/helpers/spacing_extension.dart';
-import '../../../../core/theming/text_styles.dart';
-import '../../data/models/specialization_response_model.dart';
+import '../../../../../core/helpers/spacing_extension.dart';
+import '../../../../../core/theming/text_styles.dart';
+import '../../../data/models/specialization_response_model.dart';
 
 class DoctorsListViewItem extends StatelessWidget {
   const DoctorsListViewItem({
