+++ lib/features/home/ui/widgets/speciality_list/specialty_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:doc/features/home/data/models/specialization_response_model.dart';
 import 'package:doc/features/home/logic/home_cubit.dart';
-import 'package:doc/features/home/ui/widgets/speciality_list_view_item.dart';
+import 'package:doc/features/home/ui/widgets/speciality_list/speciality_list_view_item.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
